With Inventory features and Delivery Confirmation.

+++ app/src/main/java/ph/com/gs3/loyaltystore/models/sqlite/dao/StoreDao.java
@@ -25,7 +25,7 @@ public class StoreDao extends AbstractDao<Store, Long> {
     public static class Properties {
         public final static Property Id = new Property(0, Long.class, "id", true, "_id");
         public final static Property Device_id = new Property(1, String.class, "device_id", false, "DEVICE_ID");
-        public final static Property Device_web_id = new Property(2, Integer.class, "device_web_id", false, "DEVICE_WEB_ID");
+        public final static Property Device_web_id = new Property(2, Long.class, "device_web_id", false, "DEVICE_WEB_ID");
         public final static Property Name = new Property(3, String.class, "name", false, "NAME");
         public final static Property Is_active = new Property(4, Integer.class, "is_active", false, "IS_ACTIVE");
         public final static Property Created_at = new Property(5, java.util.Date.class, "created_at", false, "CREATED_AT");
@@ -75,7 +75,7 @@ public class StoreDao extends AbstractDao<Store, Long> {
             stmt.bindString(2, device_id);
         }
  
-        Integer device_web_id = entity.getDevice_web_id();
+        Long device_web_id = entity.getDevice_web_id();
         if (device_web_id != null) {
             stmt.bindLong(3, device_web_id);
         }
@@ -113,7 +113,7 @@ public class StoreDao extends AbstractDao<Store, Long> {
         Store entity = new Store( //
             cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0), // id
             cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1), // device_id
-            cursor.isNull(offset + 2) ? null : cursor.getInt(offset + 2), // device_web_id
+            cursor.isNull(offset + 2) ? null : cursor.getLong(offset + 2), // device_web_id
             cursor.isNull(offset + 3) ? null : cursor.getString(offset + 3), // name
             cursor.isNull(offset + 4) ? null : cursor.getInt(offset + 4), // is_active
             cursor.isNull(offset + 5) ? null : new java.util.Date(cursor.getLong(offset + 5)), // created_at
@@ -127,7 +127,7 @@ public class StoreDao extends AbstractDao<Store, Long> {
     public void readEntity(Cursor cursor, Store entity, int offset) {
         entity.setId(cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0));
         entity.setDevice_id(cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1));
-        entity.setDevice_web_id(cursor.isNull(offset + 2) ? null : cursor.getInt(offset + 2));
+        entity.setDevice_web_id(cursor.isNull(offset + 2) ? null : cursor.getLong(offset + 2));
         entity.setName(cursor.isNull(offset + 3) ? null : cursor.getString(offset + 3));
         entity.setIs_active(cursor.isNull(offset + 4) ? null : cursor.getInt(offset + 4));
         entity.setCreated_at(cursor.isNull(offset + 5) ? null : new java.util.Date(cursor.getLong(offset + 5)));
